refactor: add application folder tree
