+++ lib/layers/application/common/app_color.dart
@@ -1,4 +1,4 @@
-import 'package:diary_toy_project/color_utils.dart';
+import 'package:diary_toy_project/layers/application/utils/color_utils.dart';
 
 final class AppColor {
   static final label = ColorUtils.fromHex('717171');
