--- lib/sized_box_utils.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-extension SizedBoxUtils on SizedBox {
-  static const normal = SizedBox(width: 8, height: 8);
-  static const twoX = SizedBox(width: 16, height: 16);
-  static const threeX = SizedBox(width: 24, height: 24);
-  static const fourX = SizedBox(width: 32, height: 32);
-}
